+++ packages/matter-node.js/test/session/secure/CasePairingTest.ts
@@ -7,7 +7,7 @@
 import * as assert from "assert";
 import { TlvEncryptedDataSigma2, TlvEncryptedDataSigma3, TlvSignedData } from "@project-chip/matter.js/session";
 import { ByteArray } from "@project-chip/matter.js/util";
-import { CryptoNode } from "../../../../src/crypto/CryptoNode";
+import { CryptoNode } from "../../../src/crypto/CryptoNode";
 
 const cryptoNode = new CryptoNode();
